Added our main program to get through the maze
This version may get us through the first quadrant of the maze, i.e., opening the gate and driving through it, but it does not detect the gate or use camera controlled motors to follow the white line. It assumes it is pointing straight, is in front of the gate, and that it can get through the gate in 2 seconds of full power to the motors. So we should be careful with testing and adjust the set_motor() and Sleep() values appropriately. Care will be needed while testing this.